[Silver I] Title: 쉬운 계단 수, Time: 124 ms, Memory: 14264 KB -BaekjoonHub

+++ 백준/Silver/10844. 쉬운 계단 수/쉬운 계단 수.java
@@ -5,7 +5,7 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int N = Integer.parseInt(br.readLine());
-        long[][] dp = new long[N + 1][10];
+        int[][] dp = new int[N + 1][10];
         for (int i = 1; i <= 9; i++) {
             dp[1][i] = 1;
         }
